arrumei salario como inteiro

+++ entidade/medico.py
@@ -2,7 +2,7 @@ from entidade.funcionario import Funcionario
 
 class Medico(Funcionario):
     def __init__(self, nome: str, cpf: str, idade: int, rua: str, numero: int, complemento: str,
-                 matricula: int, salario: float, crm: str):
+                 matricula: int, salario: int, crm: str):
         super().__init__(nome, cpf, idade, rua, numero, complemento, matricula, salario)
         if isinstance(crm, str):
             self.__crm = crm
